requirements.txt
Flask>=2.2.5
elevenlabs>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0